requirements.txt
django
pika
sendgrid
requests
python-dotenv